refactor(store): extract feed aggregation and downloads into leaf modules
src/store/index.ts had grown to ~830 lines mixing playback state, the episode
status/queue core, the offline downloads store, and the whole Latest Episodes
aggregation engine (a derived store plus seven helpers).

Pull the two self-contained concerns into their own files:
- src/store/feeds.ts: savedFeeds, episodeCache, episodeListLimit,
  sanitizeEpisodeListLimit, and the latestEpisodes projection + helpers.
- src/store/downloads.ts: the downloadedEpisodes store.

Both are leaf modules — they import only types/constants, never the rest of the
store layer — so there are no import cycles. index.ts re-exports them so `src/store`
stays the single import surface; no consumer import changed. The inherently coupled
playback/queue/episode-status core stays together in index.ts (down to ~510 lines),
since splitting it would force currentEpisode<->queue<->playedEpisodes cycles.

Pure code move, behavior-preserving. Gates green; verified live (plugin reloads
with the new bundle, view intact, no console errors).
